Tinymovr Arduino for firmware 2.3.x

+++ src/helpers.hpp
@@ -14,6 +14,7 @@
 #else
 #include <cstdint>
 #include <cstddef>
+#include <cstring>
 #endif
 
 #if defined ARDUINO
@@ -21,11 +22,11 @@
 #endif
 
 #define CAN_EP_SIZE (12)
-#define CAN_EP_MASK ((1 << CAN_EP_SIZE) - 1)
+#define CAN_EP_MASK ((1UL << CAN_EP_SIZE) - 1)
 #define CAN_SEQ_SIZE (9)
-#define CAN_SEQ_MASK (((1 << CAN_SEQ_SIZE) - 1) << CAN_EP_SIZE)
+#define CAN_SEQ_MASK (((1UL << CAN_SEQ_SIZE) - 1) << CAN_EP_SIZE)
 #define CAN_DEV_SIZE (8)
-#define CAN_DEV_MASK (((1 << CAN_DEV_SIZE) - 1) << (CAN_EP_SIZE + CAN_SEQ_SIZE))
+#define CAN_DEV_MASK (((1UL << CAN_DEV_SIZE) - 1) << (CAN_EP_SIZE + CAN_SEQ_SIZE))
 
 typedef void (*send_callback)(uint32_t arbitration_id, uint8_t *data, uint8_t dlc, bool rtr);
 typedef bool (*recv_callback)(uint32_t *arbitration_id, uint8_t *data, uint8_t *dlc);
@@ -47,7 +48,7 @@ class Node {
     uint8_t _dlc;
     uint32_t get_arbitration_id(uint32_t cmd_id)
     {
-        return ((this->can_node_id << (CAN_EP_SIZE + CAN_SEQ_SIZE)) & CAN_DEV_MASK) | (cmd_id & CAN_EP_MASK);
+        return ((((uint32_t)this->can_node_id) << (CAN_EP_SIZE + CAN_SEQ_SIZE)) & CAN_DEV_MASK) | (cmd_id & CAN_EP_MASK);
     }
     void send(uint32_t cmd_id, uint8_t *data, uint8_t data_size, bool rtr)
     {
